update transformer dtype

+++ python-app/src/main.py
@@ -60,7 +60,7 @@ def main():
     transformer = QwenImageTransformer2DModel.from_single_file(
         transformer_path,
         config = transformer_config_path,
-        torch_dtype=torch_dtype,
+        torch_dtype=torch.float8_e4m3fn,
         use_safetensors=True,
         local_files_only=True
     )
